More JS cleanup & reformatting
All JS should be formatted the same now.
Added new functions in as necessary to replace reud calls.
UI.JS still needs alot of work. I need to experiment with a few
functions to determine if some methods may be faster than others.
Unfortunately, my understanding on how memory is handled on this
platform is lacking so I cannot take that into account.